:recycle: LoginInput component toggle 옮김 (#5)
